requirements.txt
Django>=4.1,<4.2
wagtail
psycopg2-binary
djangorestframework
markdown
django-auditlog
django-filter
google-cloud-recaptcha-enterprise
wagtailosm
django-cors-headers
gdal
django-threadlocals
pyproj
pip
django-storages
djangorestframework-recaptcha
gunicorn
firebase_admin
firebase-rest-api
drf_dynamic_fields
whitenoise
django-session-timeout
drf-spectacular
django-defender